requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
cryptography==36.0.1
discord.py==1.7.3
idna==3.1
multidict==5.2.0
pip==21.3.1
pycparser==2.21
pyOpenSSL==21.0.0
PySocks==1.7.1
pytz==2021.3
requests==2.27.1
setuptools==60.5.0
six==1.16.0
soupsieve==2.3.1
typing_extensions==4.0.1
urllib3==1.26.8
wheel==0.37.1
yarl==1.7.2
